requirements.txt
azure-storage-blob==12.19.0
python-dotenv==1.0.0
requests==2.32.5